hackathon/requirements.txt
blinker==1.6.2
click==8.1.4
Flask==2.3.2
Flask-Reuploaded==1.3.0
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.0
google-cloud-aiplatform~=0.5.1
gunicorn==20.1.0
google-cloud-bigquery==3.11.0
google-cloud-bigquery-storage==2.19.1
#Flask==2.1.0
#Flask-Reuploaded~=1.3.0
#gunicorn==20.1.0
#google-cloud-aiplatform~=0.5.1
#Flask-WTF==0.14.3